refactor: separate notice form logic into custom hook
